+++ ReleaseNotes.md
@@ -19,6 +19,9 @@
 - Fixed issue with Display Text formatter with int/float type mismatch (for example, now {0:D6} works for having text with 6 digits, even if the number is a float - before it would just fail)
 - Promoted platformer snippet to a full sample
 - Separated Okapi Kit and the samples in two different packages
+- Added Singleton system
+  - Singletons are objects that enforce only one existing and that survive scene transitions
+- Added ChangeSound action
 
 ## V1.8.0